move ops to their own files
