+++ tfjs-core/src/ops/array_ops_test.ts
@@ -762,80 +762,6 @@ describeWithFlags('onesLike', ALL_ENVS, () => {
     expectArraysEqual(await res.data(), [1, 1, 1, 1]);
   });
 });
-
-describeWithFlags('eye', ALL_ENVS, () => {
-  it('1x1', async () => {
-    const r = tf.eye(1);
-    expectArraysClose(await r.data(), [1]);
-    expect(r.shape).toEqual([1, 1]);
-    expect(r.dtype).toBe('float32');
-  });
-
-  it('2x2', async () => {
-    const r = tf.eye(2);
-    expect(r.shape).toEqual([2, 2]);
-    expect(r.dtype).toBe('float32');
-    expectArraysClose(await r.data(), [1, 0, 0, 1]);
-  });
-
-  it('3x3', async () => {
-    const r = tf.eye(3);
-    expect(r.shape).toEqual([3, 3]);
-    expect(r.dtype).toBe('float32');
-    expectArraysClose(await r.data(), [1, 0, 0, 0, 1, 0, 0, 0, 1]);
-  });
-
-  it('3x4', async () => {
-    const r = tf.eye(3, 4);
-    expect(r.shape).toEqual([3, 4]);
-    expect(r.dtype).toBe('float32');
-    expectArraysClose(await r.data(), [1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0]);
-  });
-
-  it('4x3', async () => {
-    const r = tf.eye(4, 3);
-    expect(r.shape).toEqual([4, 3]);
-    expectArraysClose(await r.data(), [1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0]);
-  });
-
-  it('with 1D batchShape', async () => {
-    const r = tf.eye(2, 2, [3]);
-    expect(r.shape).toEqual([3, 2, 2]);
-    expectArraysClose(await r.data(), [1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1]);
-  });
-
-  it('with 2D batchShape', async () => {
-    const r = tf.eye(2, 2, [2, 3]);
-    expect(r.shape).toEqual([2, 3, 2, 2]);
-    expectArraysClose(await r.data(), [
-      1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1
-    ]);
-  });
-
-  it('with 3D batchShape', async () => {
-    const r = tf.eye(2, 2, [2, 2, 3]);
-    expect(r.shape).toEqual([2, 2, 3, 2, 2]);
-    expectArraysClose(await r.data(), [
-      1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1,
-      1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1
-    ]);
-  });
-
-  it('3x3, int32', async () => {
-    const r = tf.eye(3, 3, null, 'int32');
-    expect(r.dtype).toBe('int32');
-    expect(r.shape).toEqual([3, 3]);
-    expectArraysClose(await r.data(), [1, 0, 0, 0, 1, 0, 0, 0, 1]);
-  });
-
-  it('3x3, bool', async () => {
-    const r = tf.eye(3, 3, null, 'bool');
-    expect(r.dtype).toBe('bool');
-    expect(r.shape).toEqual([3, 3]);
-    expectArraysClose(await r.data(), [1, 0, 0, 0, 1, 0, 0, 0, 1]);
-  });
-});
-
 class MockContext {
   getImageData(x: number, y: number, width: number, height: number) {
     const data = new Uint8ClampedArray(width * height * 4);
@@ -1500,223 +1426,6 @@ describeWithFlags('clone', ALL_ENVS, () => {
   it('throws when passed a non-tensor', () => {
     expect(() => tf.clone({} as tf.Tensor))
         .toThrowError(/Argument 'x' passed to 'clone' must be a Tensor/);
-  });
-});
-
-describeWithFlags('tile', ALL_ENVS, () => {
-  it('1D (tile)', async () => {
-    const t = tf.tensor1d([1, 2, 3]);
-    const t2 = tf.tile(t, [2]);
-
-    expect(t2.shape).toEqual([6]);
-    expectArraysClose(await t2.data(), [1, 2, 3, 1, 2, 3]);
-  });
-
-  it('2D (tile)', async () => {
-    const t = tf.tensor2d([1, 11, 2, 22], [2, 2]);
-    let t2 = tf.tile(t, [1, 2]);
-
-    expect(t2.shape).toEqual([2, 4]);
-    expectArraysClose(await t2.data(), [1, 11, 1, 11, 2, 22, 2, 22]);
-
-    t2 = tf.tile(t, [2, 1]);
-    expect(t2.shape).toEqual([4, 2]);
-    expectArraysClose(await t2.data(), [1, 11, 2, 22, 1, 11, 2, 22]);
-
-    t2 = tf.tile(t, [2, 2]);
-    expect(t2.shape).toEqual([4, 4]);
-    expectArraysClose(
-        await t2.data(),
-        [1, 11, 1, 11, 2, 22, 2, 22, 1, 11, 1, 11, 2, 22, 2, 22]);
-  });
-
-  it('3D (tile)', async () => {
-    const t = tf.tensor3d([1, 2, 3, 4, 5, 6, 7, 8], [2, 2, 2]);
-    const t2 = tf.tile(t, [1, 2, 1]);
-
-    expect(t2.shape).toEqual([2, 4, 2]);
-    expectArraysClose(
-        await t2.data(), [1, 2, 3, 4, 1, 2, 3, 4, 5, 6, 7, 8, 5, 6, 7, 8]);
-  });
-
-  it('4D (tile)', async () => {
-    const t = tf.tensor4d([1, 2, 3, 4, 5, 6, 7, 8], [1, 2, 2, 2]);
-    const t2 = tf.tile(t, [1, 2, 1, 1]);
-
-    expect(t2.shape).toEqual([1, 4, 2, 2]);
-    expectArraysClose(
-        await t2.data(), [1, 2, 3, 4, 5, 6, 7, 8, 1, 2, 3, 4, 5, 6, 7, 8]);
-  });
-
-  it('5D (tile)', async () => {
-    const t = tf.tensor5d([1, 2, 3, 4, 5, 6, 7, 8], [1, 1, 2, 2, 2]);
-    const t2 = tf.tile(t, [1, 2, 1, 1, 1]);
-
-    expect(t2.shape).toEqual([1, 2, 2, 2, 2]);
-    expectArraysClose(
-        await t2.data(), [1, 2, 3, 4, 5, 6, 7, 8, 1, 2, 3, 4, 5, 6, 7, 8]);
-  });
-
-  it('1d string tensor', async () => {
-    const a = tf.tensor(['a', 'b', 'c']);
-    const res = tf.tile(a, [2]);
-    expect(res.shape).toEqual([6]);
-    expectArraysEqual(await res.data(), ['a', 'b', 'c', 'a', 'b', 'c']);
-  });
-
-  it('2d string tensor', async () => {
-    const a = tf.tensor([['a', 'b'], ['c', 'd']]);
-    const res = tf.tile(a, [2, 3]);
-    expect(res.shape).toEqual([4, 6]);
-    expectArraysEqual(await res.data(), [
-      'a', 'b', 'a', 'b', 'a', 'b', 'c', 'd', 'c', 'd', 'c', 'd',
-      'a', 'b', 'a', 'b', 'a', 'b', 'c', 'd', 'c', 'd', 'c', 'd'
-    ]);
-  });
-
-  it('propagates NaNs', async () => {
-    const t = tf.tensor1d([1, 2, NaN]);
-
-    const t2 = tf.tile(t, [2]);
-
-    expect(t2.shape).toEqual([6]);
-    expectArraysClose(await t2.data(), [1, 2, NaN, 1, 2, NaN]);
-  });
-
-  it('1D bool (tile)', async () => {
-    const t = tf.tensor1d([true, false, true], 'bool');
-    const t2 = tf.tile(t, [2]);
-
-    expect(t2.shape).toEqual([6]);
-    expect(t2.dtype).toBe('bool');
-    expectArraysEqual(await t2.data(), [1, 0, 1, 1, 0, 1]);
-  });
-
-  it('2D bool (tile)', async () => {
-    const t = tf.tensor2d([true, false, true, true], [2, 2], 'bool');
-    let t2 = tf.tile(t, [1, 2]);
-
-    expect(t2.shape).toEqual([2, 4]);
-    expect(t2.dtype).toBe('bool');
-    expectArraysEqual(await t2.data(), [1, 0, 1, 0, 1, 1, 1, 1]);
-
-    t2 = tf.tile(t, [2, 1]);
-    expect(t2.shape).toEqual([4, 2]);
-    expect(t2.dtype).toBe('bool');
-    expectArraysEqual(await t2.data(), [1, 0, 1, 1, 1, 0, 1, 1]);
-
-    t2 = tf.tile(t, [2, 2]);
-    expect(t2.shape).toEqual([4, 4]);
-    expect(t2.dtype).toBe('bool');
-    expectArraysEqual(
-        await t2.data(), [1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1]);
-  });
-
-  it('3D bool (tile)', async () => {
-    const t = tf.tensor3d(
-        [true, false, true, false, true, false, true, false], [2, 2, 2],
-        'bool');
-    const t2 = tf.tile(t, [1, 2, 1]);
-
-    expect(t2.shape).toEqual([2, 4, 2]);
-    expect(t2.dtype).toBe('bool');
-    expectArraysEqual(
-        await t2.data(), [1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0]);
-  });
-
-  it('1D int32 (tile)', async () => {
-    const t = tf.tensor1d([1, 2, 5], 'int32');
-    const t2 = tf.tile(t, [2]);
-
-    expect(t2.shape).toEqual([6]);
-    expect(t2.dtype).toBe('int32');
-    expectArraysEqual(await t2.data(), [1, 2, 5, 1, 2, 5]);
-  });
-
-  it('2D int32 (tile)', async () => {
-    const t = tf.tensor2d([1, 2, 3, 4], [2, 2], 'int32');
-    let t2 = tf.tile(t, [1, 2]);
-
-    expect(t2.shape).toEqual([2, 4]);
-    expect(t2.dtype).toBe('int32');
-    expectArraysEqual(await t2.data(), [1, 2, 1, 2, 3, 4, 3, 4]);
-
-    t2 = tf.tile(t, [2, 1]);
-    expect(t2.shape).toEqual([4, 2]);
-    expect(t2.dtype).toBe('int32');
-    expectArraysEqual(await t2.data(), [1, 2, 3, 4, 1, 2, 3, 4]);
-
-    t2 = tf.tile(t, [2, 2]);
-    expect(t2.shape).toEqual([4, 4]);
-    expect(t2.dtype).toBe('int32');
-    expectArraysEqual(
-        await t2.data(), [1, 2, 1, 2, 3, 4, 3, 4, 1, 2, 1, 2, 3, 4, 3, 4]);
-  });
-
-  it('3D int32 (tile)', async () => {
-    const t = tf.tensor3d([1, 2, 3, 4, 5, 6, 7, 8], [2, 2, 2], 'int32');
-    const t2 = tf.tile(t, [1, 2, 1]);
-
-    expect(t2.shape).toEqual([2, 4, 2]);
-    expect(t2.dtype).toBe('int32');
-    expectArraysEqual(
-        await t2.data(), [1, 2, 3, 4, 1, 2, 3, 4, 5, 6, 7, 8, 5, 6, 7, 8]);
-  });
-
-  it('1D (tile) gradient', async () => {
-    const x = tf.tensor1d([1, 2, 3]);
-    const dy = tf.tensor1d([0.1, 0.2, 0.3, 1, 2, 3, 10, 20, 30]);
-    const gradients = tf.grad(x => tf.tile(x, [3]))(x, dy);
-    expectArraysClose(await gradients.data(), [11.1, 22.2, 33.3]);
-    expect(gradients.shape).toEqual([3]);
-  });
-
-  it('gradient with clones', async () => {
-    const x = tf.tensor1d([1, 2, 3]);
-    const dy = tf.tensor1d([0.1, 0.2, 0.3, 1, 2, 3, 10, 20, 30]);
-    const gradients = tf.grad(x => tf.tile(x.clone(), [3]).clone())(x, dy);
-    expectArraysClose(await gradients.data(), [11.1, 22.2, 33.3]);
-    expect(gradients.shape).toEqual([3]);
-  });
-
-  it('2D (tile) gradient', async () => {
-    const x = tf.tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const dy = tf.tensor2d([[1, 2, 10, 20], [3, 4, 30, 40]], [2, 4]);
-    const gradients = tf.grad(x => tf.tile(x, [1, 2]))(x, dy);
-    expectArraysClose(await gradients.data(), [11, 22, 33, 44]);
-    expect(gradients.shape).toEqual([2, 2]);
-  });
-
-  it('3D (tile) gradient', async () => {
-    const x = tf.tensor3d([[[1], [2]], [[3], [4]]], [2, 2, 1]);
-    const dy = tf.tensor3d([[[1, 10], [2, 20]], [[3, 30], [4, 40]]], [2, 2, 2]);
-    const gradients = tf.grad(x => tf.tile(x, [1, 1, 2]))(x, dy);
-    expectArraysClose(await gradients.data(), [11, 22, 33, 44]);
-    expect(gradients.shape).toEqual([2, 2, 1]);
-  });
-
-  it('4D (tile) gradient', async () => {
-    const x = tf.tensor4d([[[[1]], [[2]]], [[[3]], [[4]]]], [2, 2, 1, 1]);
-    const dy = tf.tensor4d(
-        [
-          [[[.01, .1], [1, 10]], [[.02, .2], [2, 20]]],
-          [[[.03, .3], [3, 30]], [[.04, .4], [4, 40]]]
-        ],
-        [2, 2, 2, 2]);
-    const gradients = tf.grad(x => tf.tile(x, [1, 1, 2, 2]))(x, dy);
-    expectArraysClose(await gradients.data(), [11.11, 22.22, 33.33, 44.44]);
-    expect(gradients.shape).toEqual([2, 2, 1, 1]);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.tile({} as tf.Tensor, [1]))
-        .toThrowError(/Argument 'x' passed to 'tile' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const res = tf.tile([1, 2, 3], [2]);
-    expect(res.shape).toEqual([6]);
-    expectArraysClose(await res.data(), [1, 2, 3, 1, 2, 3]);
   });
 });
