+++ tfjs-core/src/ops/ops.ts
@@ -18,6 +18,7 @@
 // Modularized ops.
 export {broadcastTo} from './broadcast_to';
 export {clone} from './clone';
+export {eye} from './eye';
 export {multinomial} from './multinomial';
 export {oneHot} from './one_hot';
 export {rand} from './rand';
@@ -26,6 +27,7 @@ export {randomNormal} from './random_normal';
 export {randomUniform} from './random_uniform';
 export {square} from './square';
 export {squaredDifference} from './squared_difference';
+export {tile} from './tile';
 export {truncatedNormal} from './truncated_normal';
 
 export * from './batchnorm';
